@@ -1,13 +1,35 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.1.2" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.1.9" />
+    <Compile Remove="wwwroot\images\user\**" />
+    <Content Remove="wwwroot\images\user\**" />
+    <EmbeddedResource Remove="wwwroot\images\user\**" />
+    <None Remove="wwwroot\images\user\**" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Remove="Migrations\20200605225205_Intial.cs" />
+    <Compile Remove="Migrations\20200605225205_Intial.Designer.cs" />
+    <Compile Remove="Migrations\20200607073942_initial.cs" />
+    <Compile Remove="Migrations\20200607073942_initial.Designer.cs" />
+    <Compile Remove="Migrations\20200607080517_Initial.cs" />
+    <Compile Remove="Migrations\20200607080517_Initial.Designer.cs" />
+    <Compile Remove="Migrations\20200607081339_Initial.cs" />
+    <Compile Remove="Migrations\20200607081339_Initial.Designer.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.DataProtection.EntityFrameworkCore" Version="3.1.4" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="3.1.4" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="3.1.4">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
   </ItemGroup>
 
 </Project>
