@@ -2,31 +2,183 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using EnergyStyleCore.Models;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.EntityFrameworkCore;
+using EnergyStyleCore.Models;
+using EnergyStyleCore.Models.Classes;
+using CustomPolicyProvider;
 
 namespace EnergyStyleCore.Controllers
 {
     public class ProductsController : Controller
     {
-        DataBase db = DataBase.GetInstance();
+        private readonly AppDBContext _context;
 
-        public IActionResult Index(int categoryId = 1)
+        public ProductsController(AppDBContext context)
         {
-            ViewBag.Categories = db.Categories;
-            var products = db.Products.Where(x => x.Category.Id == categoryId);
+            _context = context;
+        }
+
+        // GET: Products
+        public async Task<IActionResult> Index(int categoryId = 1)
+        {
+            ViewBag.Categories = await _context.Categories.ToListAsync();
+            var products = await _context.Products
+                .Include(product => product.Category)
+                .Include(product => product.Characteristics)
+                .Where(x => x.Category.Id == categoryId)
+                .ToListAsync();
             return View(products);
         }
 
-        public IActionResult Details(int id)
+        public async Task<IActionResult> Sales()
         {
-            var product = db.Products.First(x => x.Id == id);
+            ViewBag.Categories = await _context.Categories.ToListAsync();
+            var products = await _context.Products
+                .Include(product => product.Category)
+                .Include(product => product.Characteristics)
+                //.Where(x => x.Sale > 0)
+                .ToListAsync();
+            return View(nameof(Index), products);
+        }
+
+        // GET: Products/Details/5
+        public async Task<IActionResult> Details(int? id)
+        {
+            if (id == null)
+                return NotFound();
+
+            var product = await _context.Products
+                .Include(product => product.Category)
+                .Include(product => product.Characteristics)
+                .FirstOrDefaultAsync(m => m.Id == id);
+            if (product == null)
+            {
+                return NotFound();
+            }
+
             return View(product);
         }
 
-        public IActionResult Sales()
+        // GET: Products/Create
+        [RoleAuthorize("user")]
+        public async Task<IActionResult> Create()
         {
+            ViewBag.Categories = await _context.Categories.ToListAsync();
             return View();
+        }
+
+        // POST: Products/Create
+        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
+        // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+        [RoleAuthorize("user")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Create(
+            [Bind("Id,Name,Price,Model,Count,Images,Category,Characteristics")] Product product)//////////////////////////////////////////////////
+        {
+            if (ModelState.IsValid)
+            {
+                var category = _context.Categories.Find(product.Category.Id);
+                product.Category = category;
+                _context.Add(product);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            return View(product);
+        }
+
+        // GET: Products/Edit/5
+        [RoleAuthorize("user")]
+        public async Task<IActionResult> Edit(int? id)
+        {
+            if (id == null)
+                return NotFound();
+
+            var product = await _context.Products
+                .Include(product => product.Category)
+                .Include(product => product.Characteristics)
+                .FirstOrDefaultAsync(m => m.Id == id);
+            if (product == null)
+            {
+                return NotFound();
+            }
+            ViewBag.Categories = await _context.Categories.ToListAsync();
+            return View(product);
+        }
+
+        // POST: Products/Edit/5
+        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
+        // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+        [RoleAuthorize("user")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Edit(int id,
+            [Bind("Id,Name,Price,Model,Count,Images,Category,Characteristics")] Product product)
+        {
+            if (id != product.Id)
+                return NotFound();
+
+            if (ModelState.IsValid)
+            {
+                try
+                {
+                    var category = _context.Categories.Find(product.Category.Id);
+                    product.Category = category;
+                    _context.Update(product);
+                    await _context.SaveChangesAsync();
+                }
+                catch (DbUpdateConcurrencyException)
+                {
+                    if (!ProductExists(product.Id))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
+                }
+                return RedirectToAction(nameof(Index));
+            }
+            return View(product);
+        }
+
+        // GET: Products/Delete/5
+        [RoleAuthorize("user")]
+        public async Task<IActionResult> Delete(int? id)
+        {
+            if (id == null)
+                return NotFound();
+
+            var product = await _context.Products
+                .Include(product => product.Category)
+                .Include(product => product.Characteristics)
+                .FirstOrDefaultAsync(m => m.Id == id);
+            if (product == null)
+            {
+                return NotFound();
+            }
+
+            return View(product);
+        }
+
+        // POST: Products/Delete/5
+        [RoleAuthorize("user")]
+        [HttpPost, ActionName("Delete")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> DeleteConfirmed(int id)
+        {
+            var product = await _context.Products.FindAsync(id);
+            _context.Products.Remove(product);
+            await _context.SaveChangesAsync();
+            return RedirectToAction(nameof(Index));
+        }
+
+        private bool ProductExists(int id)
+        {
+            return _context.Products.Any(e => e.Id == id);
         }
     }
 }
