@@ -3,14 +3,158 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.EntityFrameworkCore;
+using EnergyStyleCore.Models;
+using EnergyStyleCore.Models.Classes;
+using CustomPolicyProvider;
 
 namespace EnergyStyleCore.Controllers
 {
     public class ServicesController : Controller
     {
-        public IActionResult Index()
+        private readonly AppDBContext _context;
+
+        public ServicesController(AppDBContext context)
+        {
+            _context = context;
+        }
+
+        // GET: Services
+        public async Task<IActionResult> Index()
+        {
+            return View(await _context.Service.ToListAsync());
+        }
+
+        // GET: Services/Details/5
+        public async Task<IActionResult> Details(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var service = await _context.Service
+                .FirstOrDefaultAsync(m => m.Id == id);
+            if (service == null)
+            {
+                return NotFound();
+            }
+
+            return View(service);
+        }
+
+        // GET: Services/Create
+        [RoleAuthorize("user")]
+        public IActionResult Create()
         {
             return View();
+        }
+
+        // POST: Services/Create
+        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
+        // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+        [RoleAuthorize("user")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Create([Bind("Id,Title,Name,Images,Info")] Service service)
+        {
+            if (ModelState.IsValid)
+            {
+                _context.Add(service);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            return View(service);
+        }
+
+        // GET: Services/Edit/5
+        [RoleAuthorize("user")]
+        public async Task<IActionResult> Edit(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var service = await _context.Service.FindAsync(id);
+            if (service == null)
+            {
+                return NotFound();
+            }
+            return View(service);
+        }
+
+        // POST: Services/Edit/5
+        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
+        // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+        [RoleAuthorize("user")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Title,Name,Images,Info")] Service service)
+        {
+            if (id != service.Id)
+            {
+                return NotFound();
+            }
+
+            if (ModelState.IsValid)
+            {
+                try
+                {
+                    _context.Update(service);
+                    await _context.SaveChangesAsync();
+                }
+                catch (DbUpdateConcurrencyException)
+                {
+                    if (!ServiceExists(service.Id))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
+                }
+                return RedirectToAction(nameof(Index));
+            }
+            return View(service);
+        }
+
+        // GET: Services/Delete/5
+        [RoleAuthorize("user")]
+        public async Task<IActionResult> Delete(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
+            var service = await _context.Service
+                .FirstOrDefaultAsync(m => m.Id == id);
+            if (service == null)
+            {
+                return NotFound();
+            }
+
+            return View(service);
+        }
+
+        // POST: Services/Delete/5
+        [RoleAuthorize("user")]
+        [HttpPost, ActionName("Delete")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> DeleteConfirmed(int id)
+        {
+            var service = await _context.Service.FindAsync(id);
+            _context.Service.Remove(service);
+            await _context.SaveChangesAsync();
+            return RedirectToAction(nameof(Index));
+        }
+
+        private bool ServiceExists(int id)
+        {
+            return _context.Service.Any(e => e.Id == id);
         }
     }
 }
