@@ -1,83 +1,79 @@
 ﻿<div class="wrapper">
     <div class="about">
-        <div class="container">
-            <div class="content">
-                <div class="about__company">
+        <div class="content">
+            <div class="about__company">
 
-                    <h1>
-                        Компания «Энергостиль» предоставляет комплекс услуг по расчету, комплектации, монтажу систем отопления,
-                        горячего и холодного водоснабжения, водоотведения. Производство санитарно-технических работ.
-                    </h1>
+                <h1>
+                    Компания «Энергостиль» предоставляет комплекс услуг по расчету, комплектации, монтажу систем отопления,
+                    горячего и холодного водоснабжения, водоотведения. Производство санитарно-технических работ.
+                </h1>
 
-                    <h2>БОЛЕЕ 8 ЛЕТ ОПЫТА</h2>
+                <h2>БОЛЕЕ 8 ЛЕТ ОПЫТА</h2>
 
-                    <p>
-                        Наши клиенты – это наши друзья, которым мы всегда готовы помочь в любую минуту.
-                        Мы – это дружный и сплоченный коллектив профессионалов, любящих своё дело.
-                        Мы всегда предоставляем гарантии на наши работы.
-                    </p>
+                <p>
+                    Наши клиенты – это наши друзья, которым мы всегда готовы помочь в любую минуту.
+                    Мы – это дружный и сплоченный коллектив профессионалов, любящих своё дело.
+                    Мы всегда предоставляем гарантии на наши работы.
+                </p>
 
-                    <h2>НАШИ ГАРАНТИИ</h2>
+                <h2>НАШИ ГАРАНТИИ</h2>
 
-                    <h3>Гарантия предоставляется:</h3>
-                    <ol class="about_li" >
-                        <li>
-                            На безотказную работу систем — 5 лет (без всяких условий).
-                        </li>
-                        <li>
-                            На разъемные соединения 3 года при использовании в системе отопления незамерзающих жидкостей сертифицированных
-                            для России и 5 лет для систем отопления на воде и систем водоснабжения.
-                        </li>
-                        <li>
-                            На неразъемные соединения на весь срок службы трубопроводов, заявленный производителем труб и фитингов.
-                        </li>
-                        <li>
-                            На работу оборудования в рамках гарантийных сроков, указанных в паспортах и гарантийных талонах на оборудование.
-                            При необходимости ремонта или замены оборудования: если оборудование поставлено подрядчиком —
-                            работы производятся бесплатно если оборудование поставлено заказчиком — работы оплачиваются заказчиком.
-                        </li>
-                        <li>
-                            Гарантия на котельное оборудование распространяется только при заключении договора на обслуживание газового или
-                            жидкотопливного котла, эксплуатация твердотопливного котла такими условиями не связана и ограничена условиями
-                            гарантии производителя.
-                        </li>
-                        <li>
-                            На безотказную работу систем — 5 лет (без всяких условий).
-                        </li>
-                    </ol>
-                    <h3>Гарантия не распространяется на случаи:</h3>
-                    <ol class="about_li" >
-                        <li>
-                            Несанкционированного изменения, ремонта и иного вмешательства третьих лиц в работу смонтированных систем
-                        </li>
-                        <li>
-                            Гарантия не распространяется на случаи эксплуатации систем с нарушением ТУ и СП для материалов из которых выполнена система.
-                            То есть, если трубопроводы подверглись деформации, химическому термическому или иному виду воздействия,
-                            превышающему допустимые пределы для данного вида материалов.
-                        </li>
-                    </ol>
-                    <div class="slider">
-                        <div class="btn-infinity-left"></div>
-                        <div class="infinity-wrap size-about">
-                            <div class="infinity-slider size-about">
-                                <img src="~/images/certificate/1.BMP" alt="">
-                                <img src="~/images/certificate/2.BMP" alt="">
-                                <img src="~/images/certificate/3.jpg" alt="">
-                                <img src="~/images/certificate/4.jpg" alt="">
-                                <img src="~/images/certificate/5.jpg" alt="">
-                                <img src="~/images/certificate/6.jpg" alt="">
-                                <img src="~/images/certificate/7.jpg" alt="">
-                                <img src="~/images/certificate/8.jpg" alt="">
-                                <img src="~/images/certificate/9.jpg" alt="">
-                                <img src="~/images/certificate/10.jpg" alt="">
-                            </div>
+                <h3>Гарантия предоставляется:</h3>
+                <ol class="about_li">
+                    <li>
+                        На безотказную работу систем — 5 лет (без всяких условий).
+                    </li>
+                    <li>
+                        На разъемные соединения 3 года при использовании в системе отопления незамерзающих жидкостей сертифицированных
+                        для России и 5 лет для систем отопления на воде и систем водоснабжения.
+                    </li>
+                    <li>
+                        На неразъемные соединения на весь срок службы трубопроводов, заявленный производителем труб и фитингов.
+                    </li>
+                    <li>
+                        На работу оборудования в рамках гарантийных сроков, указанных в паспортах и гарантийных талонах на оборудование.
+                        При необходимости ремонта или замены оборудования: если оборудование поставлено подрядчиком —
+                        работы производятся бесплатно если оборудование поставлено заказчиком — работы оплачиваются заказчиком.
+                    </li>
+                    <li>
+                        Гарантия на котельное оборудование распространяется только при заключении договора на обслуживание газового или
+                        жидкотопливного котла, эксплуатация твердотопливного котла такими условиями не связана и ограничена условиями
+                        гарантии производителя.
+                    </li>
+                </ol>
+                <h3>Гарантия не распространяется на случаи:</h3>
+                <ol class="about_li">
+                    <li>
+                        Несанкционированного изменения, ремонта и иного вмешательства третьих лиц в работу смонтированных систем
+                    </li>
+                    <li>
+                        Гарантия не распространяется на случаи эксплуатации систем с нарушением ТУ и СП для материалов из которых выполнена система.
+                        То есть, если трубопроводы подверглись деформации, химическому термическому или иному виду воздействия,
+                        превышающему допустимые пределы для данного вида материалов.
+                    </li>
+                </ol>
+                <div class="slider">
+                    <div class="btn-infinity-left"></div>
+                    <div class="infinity-wrap size-about">
+                        <div class="infinity-slider size-about">
+                            <img src="~/images/certificate/1.BMP" alt="">
+                            <img src="~/images/certificate/2.BMP" alt="">
+                            <img src="~/images/certificate/3.jpg" alt="">
+                            <img src="~/images/certificate/4.jpg" alt="">
+                            <img src="~/images/certificate/5.jpg" alt="">
+                            <img src="~/images/certificate/6.jpg" alt="">
+                            <img src="~/images/certificate/7.jpg" alt="">
+                            <img src="~/images/certificate/8.jpg" alt="">
+                            <img src="~/images/certificate/9.jpg" alt="">
+                            <img src="~/images/certificate/10.jpg" alt="">
                         </div>
-                        <div class="btn-infinity-right"></div>
                     </div>
+                    <div class="btn-infinity-right"></div>
                 </div>
             </div>
         </div>
     </div>
 </div>
-
-<script src="/js/slider.js"></script>
+@section scripts{
+    <script src="~/js/slider.js"></script>
+}
