@@ -1,59 +1,71 @@
-﻿@model Product
+﻿@using System.Security.Claims
+@model EnergyStyleCore.Models.Classes.Product
+
+@{
+    ViewData["Title"] = "Details";
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
 <div class="wrapper">
     <div class="items">
-        <div class="container">
-            <div class="content">
-                <div class="items__main">
-                    <div class="items__image">
-                        <div class="slider">
-                            <div class="btn-infinity-left"></div>
-                            <div class="infinity-wrap infinity-size">
-                                <div class="infinity-slider infinity-size">
-                                    <img src="~/images/certificate/1.BMP" alt="">
-                                    <img src="~/images/certificate/2.BMP" alt="">
-                                    <img src="~/images/certificate/3.jpg" alt="">
-                                    <img src="~/images/certificate/4.jpg" alt="">
-                                    <img src="~/images/certificate/5.jpg" alt="">
-                                    <img src="~/images/certificate/6.jpg" alt="">
-                                    <img src="~/images/certificate/7.jpg" alt="">
-                                    <img src="~/images/certificate/8.jpg" alt="">
-                                    <img src="~/images/certificate/9.jpg" alt="">
-                                    <img src="~/images/certificate/10.jpg" alt="">
-                                </div>
+        <div class="content">
+            <div class="items__main">
+                <div class="items__image">
+                    <div class="slider">
+                        <div class="btn-infinity-left"></div>
+                        <div class="infinity-wrap infinity-size">
+                            <div class="infinity-slider infinity-size">
+                                @foreach (var item in Model.Images)
+                                {
+                                    <img src="@item" alt="">
+                                }
                             </div>
-                            <div class="btn-infinity-right"></div>
                         </div>
-                    </div>
-                    <div class="items__description">
-                        <div class="item__title">@Model.Name</div>
-                        <div class="item__character">
-                            <span>Модель</span>
-                            <p>@Model.Model</p>
-                            <span>Мощность</span>
-                            <p>25кВт</p>
-                        </div>
-                    </div>
-
-                    <div class="item__puhchase">
-                        <span>Цена</span>
-                        <p>@Model.Price р</p>
-                        <button class="purchaise">Купить</button>
+                        <div class="btn-infinity-right"></div>
                     </div>
                 </div>
-                <div class="item__charasteristic">
-                    <h5>Характеристики</h5>
-                    <div class="characteristic__row">
+                <div class="items__description">
+                    <div class="item__title">@Model.Name</div>
+                    <div class="item__character">
                         <span>Модель</span>
                         <p>@Model.Model</p>
-                    </div>
-
-                    <div class="characteristic__row">
-                        <span>Мощность</span>
-                        <p>24кВт</p>
+                        <span>В наличии</span>
+                        <p>@Model.Count</p>
                     </div>
                 </div>
+
+                <div class="item__puhchase">
+                    <span>Цена</span>
+                    <p>@Model.Price р</p>
+                    <button class="purchaise">Купить</button>
+                </div>
+            </div>
+            <div class="item__charasteristic">
+                <h5>Характеристики</h5>
+                @foreach (var item in Model.Characteristics)
+                {
+                    <div class="characteristic__row">
+                        <span>@item.Name</span>
+                        <p>@item.Value</p>
+                    </div>
+                }
             </div>
         </div>
     </div>
 </div>
-<script src="/js/slider.js"></script>
+
+@section Scripts {
+    <script src="/js/slider.js"></script>
+}
+@if (!string.IsNullOrWhiteSpace(User?.Identity?.Name))
+{
+    var role = User.FindFirst(c => c.Type == ClaimTypes.Role)?.Value;
+
+    if (role.Contains("user"))
+    {
+        <div>
+            <a asp-action="Edit" asp-route-id="@Model.Id">Edit</a> |
+            <a asp-action="Index">Back to List</a>
+        </div>
+    }
+}
+
