@@ -1,14 +1,14 @@
 ﻿<!DOCTYPE html>
 <html lang="en">
 <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>@ViewData["Title"] - EnergyStyleCore</title>
+    <link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.min.css" />
+    <link rel="stylesheet" href="~/css/site.css" />
     <link rel="stylesheet" href="~/css/style.css" />
     <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500&display=swap" rel="stylesheet">
     <script src="https://kit.fontawesome.com/7713828ed3.js" crossorigin="anonymous"></script>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.7/jquery.js"></script>
-
 </head>
 <body>
     <header>
@@ -19,11 +19,9 @@
                 </div>
                 <div class="header__search"><form class="search" action="/"><input type="text" placeholder="Поиск..."><button type="submit"></button></form></div>
                 <div id="header__phone" class="header__phone">
-                    <a href="/">Заказать звонок</a>
+                    <a id="call">Заказать звонок</a>
                 </div>
-                <div class="personal__enter">
-                    <a href="/"><i class="fas fa-lock"></i>Войти</a>
-                </div>
+                <partial name="_LoginPartial" />
             </div>
             <div class="header__links">
                 @Html.ActionLink(" Каталог", "Index", "Products")
@@ -33,12 +31,12 @@
                 @Html.ActionLink(" Контакты", "Contact", "Home")
             </div>
         </div>
-
     </header>
-
-    @*<partial name="_CookieConsentPartial" />*@
-
-    @RenderBody()
+    <div class="container">
+        <main role="main" class="pb-3">
+            @RenderBody()
+        </main>
+    </div>
 
     <footer class="footer">
         <div class="container">
@@ -57,21 +55,73 @@
             <div class="footer__row">Энергостиль © 2020</div>
         </div>
     </footer>
-    <script type="text/javascript">
-                    @{var url = Url.Action("GetShopsData", "Home");}
-                    $.getJSON('@url', function (data) {
-                        $.each(data, function (i, item) {
-                            let contacts = $("#contacts");
-                            contacts.append(`<p>Наши контакты</p>`);
-                            contacts.append(`<a href="tel:${item.LocalPhone}"><i class="fas fa-phone"></i>${item.LocalPhone}</a>`);
-                            contacts.append(`<a href="mailto:mail@mail.com"><i class="far fa-envelope-open"></i>${item.Email}</a>`);
-                            contacts.append(`<a href="@Url.Action("Contact", "Home")"><i class="fas fa-search-location"></i>${item.Address}</a>`);
 
-                            let header__phone = $("#header__phone");
-                            header__phone.append(`<a href="tel:${item.LocalPhone}"><i class="fas fa-phone"></i>${item.LocalPhone}</a>`);
-                        });
-                    });
+    <div id="divloading" style="margin: 0px; padding: 0px; position: fixed; right: 0px;
+        top: 0px; width: 100%; height: 100%; background-color:#ffffff; z-index: 300001;
+        opacity: .8; filter: alpha(opacity=80);display:none">
+        <div style="position: absolute; top: 40%; left: 40%; color: black;">
+            <img src="~/images/Wedges.gif" />
+        </div>
+    </div>
+
+    <div class="smodal">
+        <div class="smodal-dialog">
+            <div class="smodal-header">
+                <h3 class="smodal-title">Вопрос к Энергостиль</h3>
+                <button id="close" class="smodal-close">&times;</button>
+            </div>
+            <div class="smodal-body" style="overflow-y:scroll;">
+                <!-- вставишь сюда всё что надо -->
+                @*<form>
+                    <input value="" class="crm-webform-input" type="text" name="DEAL_TITLE" id="DEAL_TITLE" placeholder="Введите Ваше имя">
+                </form>*@
+            </div>
+        </div>
+    </div>
+
+    <script src="~/lib/jquery/dist/jquery.min.js"></script>
+    <script src="~/lib/jquery-ajax-unobtrusive/dist/jquery.unobtrusive-ajax.min.js"></script>
+    <script src="~/lib/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
+    <script src="~/js/site.js" asp-append-version="true"></script>
+    <script>
+        let $loading = $('#divloading');
+        $(document).ajaxStart(() => $loading.show());
+        $(document).ajaxStop(() => $loading.hide());
+
+        let call = document.getElementById('call');
+        let modal = document.querySelector('.smodal');
+        let modalClose = document.querySelector('.smodal-close');
+
+        function toggleModal() {
+            modal.classList.toggle('is-open');
+        }
+
+        modalClose.addEventListener('click', toggleModal);
+        call.addEventListener('click', toggleModal);
+
+        @{var url = Url.Action("GetShopsData", "Home");}
+        $.ajax({
+            url: '@url',
+            type: 'GET',
+            contentType: 'application/json',
+            success: (data) => {
+                console.log(data);
+                if (data.length > 0) {
+                    let item = data[0];
+                    let contacts = $("#contacts");
+                    contacts.append(`<p>Наши контакты</p>
+                    <a href="tel:${item.localPhone}"><i class="fas fa-phone"></i>${item.localPhone}</a>
+                    <a href="mailto:mail@mail.com"><i class="far fa-envelope-open"></i>${item.email}</a>
+                    <a href="@Url.Action("Contact", "Home")"><i class="fas fa-search-location"></i>${item.address}</a>`
+                    );
+                    let header__phone = $("#header__phone");
+                    header__phone.append(`<a href="tel:${item.localPhone}"><i class="fas fa-phone"></i>${item.localPhone}</a>`);
+                }
+            },
+            error: (data) => console.error(data),
+        });
     </script>
+
     @RenderSection("Scripts", required: false)
 </body>
 </html>
